rename some folders about view
