+++ src/view/canvas_components/accessory.ts
