+++ src/view/canvas_components/background.ts
