+++ src/view/canvas_components/canvas.ts
