+++ src/view/canvas_components/furniture.ts
