+++ src/view/canvas_components/furniture_group.ts
