+++ src/view/canvas_components/joint.ts
