+++ src/view/canvas_components/pedestal.ts
