+++ src/view/canvas_components/room.ts
