+++ src/view/canvas_components/view_factory.ts
