+++ src/view/canvas_components/view_object.ts
