+++ src/view/canvas_components/wall.ts
